Some more updates

+++ engine.py
@@ -18,7 +18,7 @@ import subprocess
 
 def edit_tcin(fin=None, fout=None, options={}, defaults={}):
     """
-    Parse a TeraChem input file.
+    Parse, modify, and/or create a TeraChem input file.
 
     Parameters
     ----------
@@ -27,7 +27,7 @@ def edit_tcin(fin=None, fout=None, options={}, defaults={}):
     fout : str, optional
         Name of the TeraChem output file to be written, if desired
     options : dict, optional
-        Dictionary of options to overrule TeraChem input file
+        Dictionary of options to overrule TeraChem input file. Pass None as value to delete a key.
     defaults : dict, optional
         Dictionary of options to add to the end
 
@@ -48,9 +48,9 @@ def edit_tcin(fin=None, fout=None, options={}, defaults={}):
             s = line.split(' ', 1)
             k = s[0].lower()
             v = s[1].strip()
-            # if k == 'coordinates':
-            #     if not os.path.exists(v.strip()):
-            #         raise RuntimeError("TeraChem coordinate file does not exist")
+            if k == 'coordinates':
+                if not os.path.exists(v.strip()):
+                    raise RuntimeError("TeraChem coordinate file does not exist")
             if k in intkeys:
                 v = int(v)
             if k in Answer:
@@ -63,11 +63,38 @@ def edit_tcin(fin=None, fout=None, options={}, defaults={}):
     for k, v in defaults.items():
         if k not in Answer.keys():
             Answer[k] = v
+    for k, v in Answer.items():
+        if v is None:
+            del Answer[k]
     # Print to the output if provided
+    havekeys = []
     if fout is not None:
         with open(fout, 'w') as f:
+            # If input file is provided, try to preserve the formatting
+            if fin is not None:
+                for line in open(fin).readlines():
+                    # Find if the line contains a key
+                    haveKey = False
+                    uncomm = line.split("#", 1)[0].strip()
+                    # Don't keep anything past the 'end' keyword
+                    if uncomm.lower() == 'end': break
+                    if len(uncomm) > 0: 
+                        haveKey = True
+                        comm = line.split("#", 1)[1].replace('\n','') if len(line.split("#", 1)) == 2 else ''
+                        s = line.split(' ', 1)
+                        w = re.findall('[ ]+',uncomm)[0]
+                        k = s[0].lower()
+                        if k in Answer:
+                            line_out = k + w + str(Answer[k]) + comm
+                            havekeys.append(k)
+                        else:
+                            line_out = line.replace('\n', '')
+                    else:
+                        line_out = line.replace('\n', '')
+                    print >> f, line_out
             for k, v in Answer.items():
-                print >> f, "%s %s" % (k, str(v))
+                if k not in havekeys:
+                    print >> f, "%-15s %s" % (k, str(v))
     return Answer
 
 def set_tcenv():
@@ -81,7 +108,6 @@ def load_tcin(f_tcin):
     tcdef = OrderedDict()
     tcdef['convthre'] = "3.0e-6"
     tcdef['threall'] = "1.0e-13"
-    tcdef['mixguess'] = "0.0"
     tcdef['scf'] = "diis+a"
     tcdef['maxit'] = "50"
     # tcdef['dftgrid'] = "1"
@@ -94,33 +120,40 @@ def load_tcin(f_tcin):
 #| Classes for external codes used  |#
 #| to calculate energy and gradient |#
 #====================================#
+stored_calcs = OrderedDict()
+
 class Engine(object):
     def __init__(self, molecule):
         if len(molecule) != 1:
             raise RuntimeError('Please pass only length-1 molecule objects to engine creation')
         self.M = deepcopy(molecule)
-        self.stored_calcs = OrderedDict()
+        # self.stored_calcs = OrderedDict()
 
-    def __deepcopy__(self, memo):
-        return copy(self)
+    # def __deepcopy__(self, memo):
+    #     return copy(self)
 
     def calc(self, coords, dirname):
-        coord_hash = tuple(list(coords))
-        if coord_hash in self.stored_calcs:
-            energy = self.stored_calcs[coord_hash]['energy']
-            gradient = self.stored_calcs[coord_hash]['gradient']
+        global stored_calcs
+        coord_hash = hash(coords.tostring())
+        if coord_hash in stored_calcs:
+            energy = stored_calcs[coord_hash]['energy']
+            gradient = stored_calcs[coord_hash]['gradient']
         else:
             if not os.path.exists(dirname): os.makedirs(dirname)
             energy, gradient = self.calc_new(coords, dirname)
-            self.stored_calcs[coord_hash] = {'coords':coords,'energy':energy,'gradient':gradient}
+            stored_calcs[coord_hash] = {'coords':coords,'energy':energy,'gradient':gradient}
         return energy, gradient
+
+    def clearCalcs(self):
+        global stored_calcs
+        stored_calcs = OrderedDict()
 
     def calc_new(self, coords, dirname):
         raise NotImplementedError("Not implemented for the base class")
 
     def calc_wq(self, coords, dirname):
-        coord_hash = tuple(list(coords))
-        if coord_hash in self.stored_calcs:
+        coord_hash = hash(coords.tostring())
+        if coord_hash in stored_calcs:
             return
         else:
             self.calc_wq_new(coords, dirname)
@@ -129,26 +162,42 @@ class Engine(object):
         raise NotImplementedError("Work Queue is not implemented for this class")
 
     def read_wq(self, coords, dirname):
-        coord_hash = tuple(list(coords))
-        if coord_hash in self.stored_calcs:
-            energy = self.stored_calcs[coord_hash]['energy']
-            gradient = self.stored_calcs[coord_hash]['gradient']
+        global stored_calcs
+        coord_hash = hash(coords.tostring())
+        if coord_hash in stored_calcs:
+            energy = stored_calcs[coord_hash]['energy']
+            gradient = stored_calcs[coord_hash]['gradient']
         else:
             if not os.path.exists(dirname):
                 raise RuntimeError("In read_wq, %s doesn't exist" % dirname)
             energy, gradient = self.read_wq_new(coords, dirname)
-            self.stored_calcs[coord_hash] = {'coords':coords,'energy':energy,'gradient':gradient}
+            stored_calcs[coord_hash] = {'coords':coords,'energy':energy,'gradient':gradient}
         return energy, gradient
 
     def read_wq_new(self, coords, dirname):
         raise NotImplementedError("Work Queue is not implemented for this class")
 
-class TeraChem(Engine):
+    def number_output(self, dirname, calcNum):
+        return
+
+class Blank(Engine):
+    """ 
+    Always return zero energy and gradient.
     """
+    def __init__(self, molecule):
+        super(Blank, self).__init__(molecule)
+
+    def calc_new(self, coords, dirname):
+        energy = 0.0
+        gradient = np.zeros(len(coords), dtype=float)
+        return energy, gradient
+
+class TeraChem(Engine):
+    """ 
     Run a TeraChem energy and gradient calculation.
     """
     def __init__(self, molecule, tcin):
-        self.tcin = tcin
+        self.tcin = tcin.copy()
         super(TeraChem, self).__init__(molecule)
 
     def calc_new(self, coords, dirname):
@@ -159,16 +208,24 @@ class TeraChem(Engine):
                 shutil.copy2(os.path.join(dirname, 'scr', f), os.path.join(dirname, f))
                 guesses.append(f)
                 have_guess = True
+        # This is for when we start geometry optimizations
+        # and we have a guess prepped and ready to go. 
         if not have_guess and 'guess' in self.tcin:
             for f in self.tcin['guess'].split():
-                shutil.copy2(f, dirname)
-                guesses.append(f)
-                have_guess = True
+                if os.path.exists(f):
+                    shutil.copy2(f, dirname)
+                    guesses.append(f)
+                    have_guess = True
+                else:
+                    del self.tcin['guess']
+                    have_guess = False
+                    break
         self.tcin['coordinates'] = 'start.xyz'
         self.tcin['run'] = 'gradient'
         if have_guess:
             self.tcin['guess'] = ' '.join(guesses)
             self.tcin['purify'] = 'no'
+            self.tcin['mixguess'] = "0.0"
         edit_tcin(fout="%s/run.in" % dirname, options=self.tcin)
         # Convert coordinates back to the xyz file
         self.M.xyzs[0] = coords.reshape(-1, 3) * 0.529177
@@ -181,6 +238,88 @@ class TeraChem(Engine):
         energy = float(open(os.path.join(dirname,'energy.txt')).readlines()[0].strip())
         gradient = np.loadtxt(os.path.join(dirname,'grad.txt')).flatten()
         return energy, gradient
+
+    def calc_wq_new(self, coords, dirname):
+        # Run TeraChem
+        wq = getWorkQueue()
+        if not os.path.exists(dirname): os.makedirs(dirname)
+        scrdir = os.path.join(dirname, 'scr')
+        if not os.path.exists(scrdir): os.makedirs(scrdir)
+        guesses = []
+        have_guess = False
+        unrestricted = self.tcin['method'][0] == 'u'
+        if unrestricted:
+            guessfnms = ['ca0', 'cb0']
+        else:
+            guessfnms = ['c0']
+        for f in ['c0', 'ca0', 'cb0']:
+            if f not in guessfnms: continue
+            if os.path.exists(os.path.join(dirname, 'scr', f)):
+                shutil.move(os.path.join(dirname, 'scr', f), os.path.join(dirname, f))
+                guesses.append(f)
+            if os.path.exists(os.path.join(dirname, f)):
+                if f not in guesses: guesses.append(f)
+        # Check if all the appropriate guess files have been found
+        # and moved to "dirname"
+        have_guess = (guesses == guessfnms)
+        # This is for when we start geometry optimizations
+        # and we have a guess prepped and ready to go. 
+        if not have_guess and 'guess' in self.tcin:
+            for f in self.tcin['guess'].split():
+                if os.path.exists(f):
+                    shutil.copy2(f, dirname)
+                    guesses.append(f)
+                    have_guess = True
+                else:
+                    del self.tcin['guess']
+                    have_guess = False
+                    break
+        self.tcin['coordinates'] = 'start.xyz'
+        self.tcin['run'] = 'gradient'
+        # For queueing up jobs, delete GPU key and let the worker decide
+        self.tcin['gpus'] = None
+        if have_guess:
+            self.tcin['guess'] = ' '.join(guesses)
+            self.tcin['purify'] = 'no'
+            self.tcin['mixguess'] = "0.0"
+        tcopts = edit_tcin(fout="%s/run.in" % dirname, options=self.tcin)
+        # Convert coordinates back to the xyz file
+        self.M.xyzs[0] = coords.reshape(-1, 3) * 0.529177
+        self.M[0].write(os.path.join(dirname, 'start.xyz'))
+        in_files = [('%s/run.in' % dirname, 'run.in'), ('%s/start.xyz' % dirname, 'start.xyz')]
+        out_files = [('%s/run.out' % dirname, 'run.out')]
+        if have_guess:
+            for g in guesses:
+                in_files.append((os.path.join(dirname, g), g))
+        for g in guessfnms:
+            out_files.append((os.path.join(dirname, 'scr', g), os.path.join('scr', g)))
+        queue_up_src_dest(wq, "%s/runtc run.in &> run.out" % rootdir, in_files, out_files, verbose=False)
+
+    def number_output(self, dirname, calcNum):
+        if not os.path.exists(os.path.join(dirname, 'run.out')):
+            raise RuntimeError('run.out does not exist')
+        shutil.copy2(os.path.join(dirname,'start.xyz'), os.path.join(dirname,'start_%03i.xyz' % calcNum))
+        shutil.copy2(os.path.join(dirname,'run.out'), os.path.join(dirname,'run_%03i.out' % calcNum))
+
+    def read_wq_new(self, coords, dirname):
+        # Extract energy and gradient
+        subprocess.call("awk '/FINAL ENERGY/ {print $3}' run.out > energy.txt", cwd=dirname, shell=True)
+        subprocess.call("awk '/Gradient units are Hartree/,/Net gradient/ {if ($1 ~ /^-?[0-9]/) {print}}' run.out > grad.txt", cwd=dirname, shell=True)
+        energy = float(open(os.path.join(dirname,'energy.txt')).readlines()[0].strip())
+        gradient = np.loadtxt(os.path.join(dirname,'grad.txt')).flatten()
+        return energy, gradient
+
+Psi4Temp = """
+molecule {{
+{charge} {mult}
+{geometry}
+}}
+
+set basis {basis}
+
+gradient('{method}{dftd}')
+
+"""
 
 class Psi4(Engine):
     """
@@ -303,6 +442,17 @@ class Psi4(Engine):
 class QChem(Engine):
     def __init__(self, molecule):
         super(QChem, self).__init__(molecule)
+        self.qcdir = False
+        self.threads = None
+
+    def nt(self):
+        if self.threads is not None:
+            return " -nt %i" % self.threads 
+        else:
+            return ""
+
+    def set_nt(self, threads):
+        self.threads = threads
 
     def calc_new(self, coords, dirname):
         if not os.path.exists(dirname): os.makedirs(dirname)
@@ -311,7 +461,13 @@ class QChem(Engine):
         self.M.edit_qcrems({'jobtype':'force'})
         self.M[0].write(os.path.join(dirname, 'run.in'))
         # Run Qchem
-        subprocess.call('%s/runqc run.in run.out &> run.log' % rootdir, cwd=dirname, shell=True)
+        if self.qcdir:
+            subprocess.call('qchem%s run.in run.out run.d &> run.log' % self.nt(), cwd=dirname, shell=True)
+        else:
+            subprocess.call('qchem%s run.in run.out run.d &> run.log' % self.nt(), cwd=dirname, shell=True)
+            # Assume reading the SCF guess is desirable
+            self.qcdir = True
+            self.M.edit_qcrems({'scf_guess':'read'})
         M1 = Molecule('%s/run.out' % dirname)
         energy = M1.qm_energies[0]
         gradient = M1.qm_grads[0].flatten()
@@ -326,7 +482,14 @@ class QChem(Engine):
         self.M[0].write(os.path.join(dirname, 'run.in'))
         in_files = [('%s/run.in' % dirname, 'run.in')]
         out_files = [('%s/run.out' % dirname, 'run.out'), ('%s/run.log' % dirname, 'run.log')]
-        queue_up_src_dest(wq, "%s/runqc run.in run.out &> run.log" % rootdir, in_files, out_files, verbose=False)
+        if self.qcdir:
+            raise RuntimeError("--qcdir currently not supported with Work Queue")
+        queue_up_src_dest(wq, "qchem%s run.in run.out &> run.log" % self.nt(), in_files, out_files, verbose=False)
+
+    def number_output(self, dirname, calcNum):
+        if not os.path.exists(os.path.join(dirname, 'run.out')):
+            raise RuntimeError('run.out does not exist')
+        shutil.copy2(os.path.join(dirname,'run.out'), os.path.join(dirname,'run_%03i.out' % calcNum))
 
     def read_wq_new(self, coords, dirname):
         M1 = Molecule('%s/run.out' % dirname)
